backend-fastapi/data-ingestion-service/requirements.txt
docx2txt
fastapi
uvicorn
openai
langchain
langchain_community
langchain_postgres
langchain-experimental
langchain-openai
pydantic
pytest
python-dotenv
psycopg[binary]
numpy
pandas
